remove memory leak because new camera are created but current camera are not deleted There is a lot of doc missing to use that class (memory management, we don't know who should delete the object).
I made a change on charedCamera value when we set the camera. However, it makes tulip really unstable (segfault after several zoom in zoom out)
I commit the modification because we can't let that and we need to clarify all the code that use the setCamera functions. 

If you need a more stable tulip, use the previous version of GlLayer (with memory leaks).



git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@3555 f5d4c5f9-c943-4916-8a57-bd8605bf961c